+++ api/src/main/java/shop/ink3/api/review/review/dto/ReviewUpdateRequest.java
@@ -1,4 +1,4 @@
-package shop.ink3.api.review.dto;
+package shop.ink3.api.review.review.dto;
 
 public record ReviewUpdateRequest(
     String title,
