+++ api/src/test/java/shop/ink3/api/review/controller/ReviewControllerTest.java
@@ -15,10 +15,6 @@ import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.boot.test.mock.mockito.MockBean;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.PageImpl;
-import org.springframework.data.domain.PageRequest;
 import org.springframework.http.MediaType;
 import org.springframework.test.context.bean.override.mockito.MockitoBean;
 import org.springframework.test.web.servlet.MockMvc;
@@ -30,10 +26,11 @@ import shop.ink3.api.book.book.entity.BookStatus;
 import shop.ink3.api.book.publisher.entity.Publisher;
 import shop.ink3.api.common.dto.PageResponse;
 import shop.ink3.api.order.orderBook.entity.OrderBook;
-import shop.ink3.api.review.dto.ReviewRequest;
-import shop.ink3.api.review.dto.ReviewResponse;
-import shop.ink3.api.review.dto.ReviewUpdateRequest;
-import shop.ink3.api.review.service.ReviewService;
+import shop.ink3.api.review.review.controller.ReviewController;
+import shop.ink3.api.review.review.dto.ReviewRequest;
+import shop.ink3.api.review.review.dto.ReviewResponse;
+import shop.ink3.api.review.review.dto.ReviewUpdateRequest;
+import shop.ink3.api.review.review.service.ReviewService;
 import shop.ink3.api.user.user.entity.User;
 import shop.ink3.api.user.user.entity.UserStatus;
