+++ api/src/test/java/shop/ink3/api/review/repository/ReviewRepositoryTest.java
@@ -21,7 +21,8 @@ import shop.ink3.api.book.book.entity.Book;
 import shop.ink3.api.book.book.entity.BookStatus;
 import shop.ink3.api.book.publisher.entity.Publisher;
 import shop.ink3.api.order.orderBook.entity.OrderBook;
-import shop.ink3.api.review.entity.Review;
+import shop.ink3.api.review.review.entity.Review;
+import shop.ink3.api.review.review.repository.ReviewRepository;
 import shop.ink3.api.user.membership.entity.Membership;
 import shop.ink3.api.user.user.entity.User;
 import shop.ink3.api.user.user.entity.UserStatus;
