+++ api/src/test/java/shop/ink3/api/review/service/ReviewServiceTest.java
@@ -27,12 +27,13 @@ import shop.ink3.api.book.publisher.entity.Publisher;
 import shop.ink3.api.common.dto.PageResponse;
 import shop.ink3.api.order.orderBook.entity.OrderBook;
 import shop.ink3.api.order.orderBook.repository.OrderBookRepository;
-import shop.ink3.api.review.dto.ReviewUpdateRequest;
-import shop.ink3.api.review.exception.ReviewNotFoundException;
-import shop.ink3.api.review.dto.ReviewRequest;
-import shop.ink3.api.review.dto.ReviewResponse;
-import shop.ink3.api.review.entity.Review;
-import shop.ink3.api.review.repository.ReviewRepository;
+import shop.ink3.api.review.review.dto.ReviewUpdateRequest;
+import shop.ink3.api.review.review.exception.ReviewNotFoundException;
+import shop.ink3.api.review.review.dto.ReviewRequest;
+import shop.ink3.api.review.review.dto.ReviewResponse;
+import shop.ink3.api.review.review.entity.Review;
+import shop.ink3.api.review.review.repository.ReviewRepository;
+import shop.ink3.api.review.review.service.ReviewService;
 import shop.ink3.api.user.user.entity.User;
 import shop.ink3.api.user.user.entity.UserStatus;
 import shop.ink3.api.user.user.repository.UserRepository;
